+++ mosqlient/registry/__init__.py
@@ -1,1 +1,2 @@
-from .models import Model #  noqa
+from .models import Model  # noqa
+from .predictions import Prediction  # noqa